+++ src/extension.ts
@@ -3,6 +3,9 @@ import * as vscode from 'vscode';
 import { LeafUiManager } from './leaf/leafAssist';
 import { LeafManager } from './leaf/leafCore';
 import { LegatoUiManager } from './legato/legatoAssist';
+import { LeafPackagesDataProvider } from './leaf/leafPackages';
+import { LeafRemotesDataProvider } from './leaf/leafRemotes';
+import { IDS } from './identifiers';
 
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
@@ -12,7 +15,7 @@ export function activate(context: vscode.ExtensionContext) {
     console.log('Extension "Legato Plugin" is now active!');
 
     // Register LeafManage to stop file watching on deactivate
-    context.subscriptions.push(LeafManager.getInstance());
+    context.subscriptions.push(LeafManager.INSTANCE);
 
     // Start Leaf UI
     new LeafUiManager().start(context);
@@ -21,6 +24,9 @@ export function activate(context: vscode.ExtensionContext) {
     // Exclude leaf-data from file watcher
     let config = vscode.workspace.getConfiguration(undefined, null);
     config.update("files.watcherExclude", { "**/leaf-data/**": true }, vscode.ConfigurationTarget.Global);
+
+	vscode.window.registerTreeDataProvider(IDS.VIEWS.PACKAGES, new LeafPackagesDataProvider(context));
+	vscode.window.registerTreeDataProvider(IDS.VIEWS.REMOTES, new LeafRemotesDataProvider(context));
 }
 
 // this method is called when your extension is deactivated