+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 - Active definition file selection
 - Basic editing support (colors) for Legato files (\*.api/\*.adef/\*.cdef/\*.mdef/\*.sdef)
 - Basic snippets Legato files
+- Add Leaf views for packages and remotes
+- Add action to configure packages in profile (from leaf package view)
+- Add actions to enable/disable remotes (from leaf remote view)
 
 ## 0.1.0
 - Initial release